+++ src/workspace/schematics/collections.ts
@@ -110,7 +110,7 @@ export class Collections {
         if (existingCollections.length > 0) {
             Output.logInfo(`${existingCollections.length} installed collection(s) detected: ${existingCollections.map((collection) => collection.name).join(', ')}`);
         } else {
-            Output.logError(`No collection found. "${defaultAngularCollection}" should be present in a correctly installed Angular CLI project.`);
+            Output.logError(`No collection found. "${defaultAngularCollection}" should be present in a correctly installed Angular CLI project. If you are in a non-Angular CLI project, try to run in the Terminal: "npm install ${defaultAngularCollection} --save-dev"`);
         }
         
         /* `.filter()` is not possible here as there is an async operation */
@@ -129,8 +129,8 @@ export class Collections {
                     watchers.push(watcher);
                 }
 
-            } catch {
-                Output.logError(`Loading of "${existingCollection.name}" collection failed.`);
+            } catch (error) {
+                Output.logError((error as Error).message);
             }
 
         }